Move around `modify_grindstone` related methods.
