+++ src/main/java/io/github/edwinmindcraft/apoli/common/power/configuration/ModifyGrindstoneConfiguration.java
@@ -8,8 +8,13 @@ import io.github.edwinmindcraft.apoli.api.IDynamicFeatureConfiguration;
 import io.github.edwinmindcraft.apoli.api.power.configuration.*;
 import io.github.edwinmindcraft.apoli.common.power.ModifyGrindstonePower;
 import io.github.edwinmindcraft.calio.api.network.CalioCodecHelper;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.core.Holder;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import org.apache.commons.lang3.mutable.Mutable;
 
 import java.util.Optional;
 
@@ -19,6 +24,7 @@ public record ModifyGrindstoneConfiguration(Holder<ConfiguredItemCondition<?, ?>
 											Holder<ConfiguredItemAction<?, ?>> lateItemAction, Holder<ConfiguredEntityAction<?, ?>> entityAction,
 											Holder<ConfiguredBlockAction<?, ?>> blockAction, ModifyGrindstonePower.ResultType resultType,
 											Optional<ConfiguredModifier<?>> experienceModifier) implements IDynamicFeatureConfiguration {
+
 	public static final Codec<ModifyGrindstoneConfiguration> CODEC = RecordCodecBuilder.create(instance -> instance.group(
 			ConfiguredItemCondition.optional("top_condition").forGetter(ModifyGrindstoneConfiguration::topItemCondition),
 			ConfiguredItemCondition.optional("bottom_condition").forGetter(ModifyGrindstoneConfiguration::bottomItemCondition),
@@ -32,5 +38,19 @@ public record ModifyGrindstoneConfiguration(Holder<ConfiguredItemCondition<?, ?>
 			CalioCodecHelper.optionalField(SerializableDataType.enumValue(ModifyGrindstonePower.ResultType.class), "result_type", ModifyGrindstonePower.ResultType.UNCHANGED).forGetter(ModifyGrindstoneConfiguration::resultType),
 			CalioCodecHelper.optionalField(ConfiguredModifier.CODEC, "xp_modifier").forGetter(ModifyGrindstoneConfiguration::experienceModifier)
 	).apply(instance, ModifyGrindstoneConfiguration::new));
+
+
+	public void tryExecute(ConfiguredPower<ModifyGrindstoneConfiguration, ModifyGrindstonePower> configuration, Entity entity, Mutable<ItemStack> itemStack, Optional<BlockPos> pos) {
+		ConfiguredItemAction.execute(configuration.getConfiguration().lateItemAction(), entity.level, itemStack);
+		ConfiguredEntityAction.execute(configuration.getConfiguration().entityAction(), entity);
+		pos.ifPresent(blockPos -> ConfiguredBlockAction.execute(configuration.getConfiguration().blockAction(), entity.level, blockPos, Direction.UP));
+	}
+
+	public boolean doesApply(ConfiguredPower<ModifyGrindstoneConfiguration, ModifyGrindstonePower> configuration, Level level, ItemStack top, ItemStack bottom, ItemStack original, Optional<BlockPos> pos) {
+		return ConfiguredItemCondition.check(configuration.getConfiguration().topItemCondition(), level, top) &&
+				ConfiguredItemCondition.check(configuration.getConfiguration().bottomItemCondition(), level, bottom) &&
+				ConfiguredItemCondition.check(configuration.getConfiguration().outputItemCondition(), level, original) &&
+				(pos.isEmpty() || ConfiguredBlockCondition.check(configuration.getConfiguration().blockCondition(), level, pos.get()));
+	}
 
 }
